backend/requirements.txt
Django==5.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
dj-database-url==2.2.0
drf-spectacular==0.27.2
groq==0.9.0
Pillow==10.4.0
whitenoise==6.7.0
gunicorn==22.0.0
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
flake8==7.1.0
black==24.8.0
